Refac actions
