+++ src/lib/y2partitioner/actions/delete_device.rb
@@ -1,4 +1,4 @@
-# Copyright (c) [2017-2019] SUSE LLC
+# Copyright (c) [2017-2020] SUSE LLC
 #
 # All Rights Reserved.
 #
@@ -24,8 +24,6 @@ require "y2partitioner/device_graphs"
 require "y2partitioner/confirm_recursive_delete"
 require "y2partitioner/immediate_unmount"
 require "y2partitioner/actions/base"
-require "y2partitioner/actions/controllers/blk_device"
-require "y2storage/filesystems/btrfs"
 require "abstract_method"
 
 module Y2Partitioner
@@ -34,7 +32,6 @@ module Y2Partitioner
     class DeleteDevice < Base
       include Yast::Logger
       include Yast::UIShortcuts
-      include ConfirmRecursiveDelete
       include ImmediateUnmount
 
       # Constructor
@@ -53,10 +50,32 @@ module Y2Partitioner
       # @return [Y2Storage::Device] device to delete
       attr_reader :device
 
-      # Deletes the indicated device
+      # Deletes the device
       #
-      # Derived classes should implement this method.
+      # Derived classes must implement this method.
+      #
+      # @see #perform_action
       abstract_method :delete
+
+      # Whether it is necessary to try to unmount (i.e., when deleting a mounted device that exists on
+      # the system)
+      #
+      # Derived classes must implement this method.
+      #
+      # @see #try_unmount
+      #
+      # @return [Boolean]
+      abstract_method :try_unmount?
+
+      # Device taken from the system devicegraph
+      #
+      # Derived classes should implement this method, although this is only required when the device can
+      # be mounted.
+      #
+      # @see #try_unmount
+      #
+      # @return [Y2Storage::Device, nil] nil if the device does not exist on disk yet.
+      abstract_method :committed_device
 
       # Checks whether delete action can be performed and if so, a confirmation popup is shown.
       # It only asks for unmounting the device it is currently mounted in the system.
@@ -74,98 +93,18 @@ module Y2Partitioner
       def perform_action
         delete
         Y2Storage::Filesystems::Btrfs.refresh_subvolumes_shadowing(device_graph)
-      end
 
-      # Current devicegraph
-      #
-      # @return [Y2Storage::Devicegraph]
-      def device_graph
-        DeviceGraphs.instance.current
+        :finish
       end
 
       # Confirmation before performing the delete action
       #
       # @return [Boolean]
       def confirm
-        if device.respond_to?(:partitions) && device.partitions.any?
-          confirm_for_partitions
-        elsif device.respond_to?(:component_of) && device.component_of.any?
-          confirm_for_component
-        else
-          simple_confirm
-        end
-      end
+        # TRANSLATORS %s is the name of the device (e.g., /dev/sda1)
+        text = format(_("Really delete %s?"), device.display_name)
 
-      # Simple confirmation to display when only the device itself is affected
-      #
-      # @see #confirm
-      #
-      # @return [Boolean]
-      def simple_confirm
-        result = Yast2::Popup.show(simple_confirm_text, buttons: :yes_no)
-        result == :yes
-      end
-
-      # Confirmation to display when the device is part of another one(s)
-      #
-      # @see #confirm
-      #
-      # @return [Boolean]
-      def confirm_for_component
-        confirm_recursive_delete(
-          device,
-          _("Confirm Deleting of Devices"),
-          _("The selected device is used by other devices in the system.\n" \
-            "To keep the system in a consistent state, the following devices\n" \
-            "will also be deleted:"),
-          recursive_confirm_text_below
-        )
-      end
-
-      # Confirmation to display when the device contains partitions
-      #
-      # @see #confirm
-      #
-      # @return [Boolean]
-      def confirm_for_partitions
-        confirm_recursive_delete(
-          device,
-          _("Confirm Deleting Device with Partitions"),
-          _("The selected device contains partitions.\n" \
-            "To keep the system in a consistent state, the following partitions\n" \
-            "and its associated devices will also be deleted:"),
-          recursive_confirm_text_below
-        )
-      end
-
-      # Text to display as final question, below the list of devices, when
-      # {#confirm_recursive_delete} is called
-      #
-      # @see #confirm_for_partitions
-      # @see #confirm_for_component
-      #
-      # @return [String]
-      def recursive_confirm_text_below
-        # TRANSLATORS %s is a kernel name like /dev/sda1
-        format(_("Really delete %s and all the affected devices?"), device.display_name)
-      end
-
-      # Text to display in {#simple_confirm}
-      #
-      # @return [String]
-      def simple_confirm_text
-        # TRANSLATORS %s is the kernel name of the device (e.g., /dev/sda1)
-        format(_("Really delete %s?"), device.display_name)
-      end
-
-      # Controller for a block device
-      #
-      # @return [Y2Partitioner::Actions::Controllers::BlkDevice, nil] nil when the device
-      #   is not a block device.
-      def controller
-        return nil unless device.is?(:blk_device)
-
-        @controller ||= Controllers::BlkDevice.new(device)
+        Yast2::Popup.show(text, buttons: :yes_no) == :yes
       end
 
       # Tries to unmount the device, if it is required.
@@ -175,7 +114,7 @@ module Y2Partitioner
       # @return [Boolean] true if it is not required to unmount or the device was correctly
       #   unmounted or the user decided to continue without unmounting; false when user cancels.
       def try_unmount
-        return true unless need_try_unmount?
+        return true unless try_unmount?
 
         # TRANSLATORS: Note added to the dialog for trying to unmount a device
         note = _("It cannot be deleted while mounted.")
@@ -183,29 +122,11 @@ module Y2Partitioner
         immediate_unmount(committed_device, note: note)
       end
 
-      # Whether it is necessary to try unmount (i.e., when deleting a mounted block device that
-      # exists on the system)
+      # Current devicegraph
       #
-      # @return [Boolean]
-      def need_try_unmount?
-        return false unless device.is?(:blk_device, :blk_filesystem)
-
-        committed_device_mounted?
-      end
-
-      # Device taken from the system devicegraph
-      #
-      # @return [Y2Storage::Device, nil] nil if the device does not exist on disk yet.
-      def committed_device
-        controller.committed_device
-      end
-
-      # Whether {#committed_device} exists and is mounted, according to the
-      # system devicegraph
-      #
-      # @return [Boolean]
-      def committed_device_mounted?
-        controller.mounted_committed_filesystem?
+      # @return [Y2Storage::Devicegraph]
+      def device_graph
+        DeviceGraphs.instance.current
       end
     end
   end
