+++ test/y2partitioner/actions/import_mount_points_test.rb
@@ -1,5 +1,6 @@
 #!/usr/bin/env rspec
-# Copyright (c) [2018] SUSE LLC
+
+# Copyright (c) [2018-2020] SUSE LLC
 #
 # All Rights Reserved.
 #
@@ -71,8 +72,8 @@ describe Y2Partitioner::Actions::ImportMountPoints do
       context "and the dialog is not accepted" do
         let(:result) { :cancel }
 
-        it "returns the result of the dialog" do
-          expect(subject.run).to eq(:cancel)
+        it "returns nil" do
+          expect(subject.run).to be_nil
         end
 
         it "does not import mount points" do
